// types/theme.d.ts
type OneOrArray<T> = T | T[]
type ObjectOf<T> = { [key: string]: T }

// ========================================================
// Props
// ========================================================

export type IProps = ObjectOf<any>

// ========================================================
// Variables
// ========================================================

export interface ISiteVariables {
  [key: string]: any

  brand: string
  htmlFontSize: string
}

export type ComponentVariableValue = any

export type ComponentVariablesObject = ObjectOf<ComponentVariableValue>

export type ComponentVariablesFunction = (siteVariables: ISiteVariables) => ComponentVariablesObject

export type ComponentVariables = ComponentVariablesObject | ComponentVariablesFunction
// ...
export interface IComponentStyles {
  [part: string]: ComponentPartStyle

  root: ComponentPartStyle
}
// ...
export type StaticStyleRenderable = string | StaticStyleObject

export type StaticStyleFunction = (siteVariables: ISiteVariables) => StaticStyleRenderable

export type StaticStyle = StaticStyleRenderable | StaticStyleFunction

export type StaticStyles = OneOrArray<StaticStyle>

// ========================================================
// Theme
// ========================================================

export interface ITheme {
  siteVariables: ISiteVariables
  componentVariables: IThemeComponentVariables
  componentStyles: IThemeComponentStyles
  rtl: boolean
}
// ...
export interface IMergedThemes {
  siteVariables: ISiteVariables
  componentVariables: IThemeComponentVariables[]
  componentStyles: IThemeComponentStyles[]
  rtl: boolean
}

export interface IThemeComponentStyles {
  [key: string]: any

  Accordion?: IComponentStyles
  Avatar?: IComponentStyles
  Button?: IComponentStyles
  Chat?: IComponentStyles
  Divider?: IComponentStyles
  Header?: IComponentStyles
  Icon?: IComponentStyles
  Image?: IComponentStyles
  Input?: IComponentStyles
  Label?: IComponentStyles
  Layout?: IComponentStyles
  List?: IComponentStyles
  Menu?: IComponentStyles
  Text?: IComponentStyles
}

export interface IThemeComponentVariables {
  [key: string]: any

  Accordion?: ComponentVariables
  Avatar?: ComponentVariables
  Button?: ComponentVariables
  Chat?: ComponentVariables
  Divider?: ComponentVariables
  Header?: ComponentVariables
  Icon?: ComponentVariables
  Image?: ComponentVariables
  Input?: ComponentVariables
  Label?: ComponentVariables
  Layout?: ComponentVariables
  List?: ComponentVariables
  Menu?: ComponentVariables
  Text?: ComponentVariables
}

// ========================================================
// Fonts
// ========================================================

export interface IFontFaceStyle {
  fontStretch?: string
  fontStyle?: string
  fontVariant?: string
  fontWeight?: number
  localAlias?: string
  unicodeRange?: string
}

export interface IFontFace {
  name: string
  paths: string[]
  style: IFontFaceStyle
}

export type FontFaces = IFontFace[]